major bug fixes
