+++ scripts/test.py
@@ -5,6 +5,7 @@ from ras_bt_framework.managers.primitive_action_manager import PrimitiveActionMa
 from ras_bt_framework.behaviors.keywords import TargetPoseMap, rotate, gripper
 from ras_bt_framework.behavior_template.keyword import keyword2module,BehaviorModule,BehaviorModuleSequence
 from ras_bt_framework.behaviors.functions import Hello,SaySomethingPy
+from ras_bt_framework.behaviors.modules import SaySomething,ThinkSomethingToSay
 from ras_bt_framework.managers.BaTMan import BaTMan
 import rclpy.node
 def hello():
@@ -24,19 +25,22 @@ def main():
         # keyword2module(tpm.move2pose_module,"move2pose",{"pose":"pose1"}),
         # keyword2module(rotate,"rotate",{"angle":90}),
         # keyword2module(gripper,"gripper",True),
+        # SaySomething(input_ports={"message":"hello"}),
         Hello(),
-        SaySomethingPy(to_say="hellopy",next_line="this is a test")
+        SaySomethingPy(to_say="hellopy",next_line="this is a test"),
+        # SaySomething(input_ports={"message":"bye"}),
         # keyword2module(hello)
     ])
     rclpy.init()
-    pam = PrimitiveActionManager(rclpy.node.Node("primitive_action_manager"))
+    # node = rclpy.node.Node("my_node")
+    # pam = PrimitiveActionManager()
     # my_generator = BehaviorTreeGenerator(pam)
     # my_generator.feed_root(myBehavior)
     # my_generator.verify_sanity()
     # my_generator.generate_xml_trees("test.xml")
     btm = BaTMan(mode_sim=True)
     btm.run_module(myBehavior,"test.xml")
-
+    rclpy.spin(btm)
 
 if __name__ == "__main__":
     main()
